Gloss

Description

Módulo para a realização da interface visual através do gloss do projeto de LI1 em 2022/23.

Documentation

type World = (Menu, Jogo, Int, ShowTime, [Picture], Jogada)

O novo tipo World define o estado Gloss

data Menu

Constructors

Opcoes Opcao

ModoJogo

PerdeuJogo

PausaJogo

data Opcao

Constructors

Jogar

Sair

type ShowTime = Float

window :: Display

background :: Color

fr :: Int

estadoGlossInicial :: [Picture] -> World

O estadoGlossInicial define o estado do Jogo ao ser inicializado a partir das texturas anteriomente dadas.

side :: Float

desenhaEstadoGloss :: World -> Picture

A função desenhaEstadoGloss toma a informação do estado do Jogo e devolve uma imagem no ecrã.

drawPlayer :: Float -> Jogador -> Jogada -> [Picture] -> Picture

A funão drawPlayer devolve a imagem do Jogador tendo em conta a direção em que está virado.

drawMapa :: Float -> Mapa -> [Picture] -> Picture

A função drawMapa desenha o mapa do jogo.

drawLinha :: Float -> Terreno -> [Obstaculo] -> [Picture] -> Picture

A função drawLinha desenha individualmente cada linha do mapa tendo em conta o tipo de terreno.

drawObstaculoRelva :: Float -> Obstaculo -> [Picture] -> Picture

A função drawObstaculoRelva desenha os obstáculos de uma linha do tipo relva

drawObstaculoRio :: Float -> Obstaculo -> [Picture] -> Picture

A função drawObstaculoRio desenha os obstáculos de uma linha do tipo rio

drawObstaculoEstrada :: Float -> Obstaculo -> Terreno -> [Picture] -> Picture

A função drawObstaculoEstrada desenha os obstáculos de uma linha do tipo estrada

mostraTempo :: ShowTime -> String

controlos :: Event -> World -> World

A função controlos altera o estado do gloss com o pressionar de certas teclas.

reageTempoGloss :: Float -> World -> World

A função reageTempoGloss altera o estado do Gloss a cada frame que passa.

estendeMapaGloss :: World -> World

A função estendeMapa verifica se as condições para que o mapa seja estendido estão cumpridas criando uma nova linha em cima do mapa e eliminando a de baixo

main :: IO ()