Tarefa1_2022li1g091

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2022/23.

Documentation

mapaValido :: Mapa -> Bool

A função mapaValido é a função principal deste documento que verifica se um dado mapa é valido através de diversas funções auxiliares

Exemplo de utilização

>>> mapaValido (Mapa 3 [(Rio 2, [Nenhum,Tronco,Tronco]),(Rio -2 ,[Nenhum,Tronco,Tronco]),(Relva, [Nenhum,Nenhum,Nenhum])])
True
>>> mapaValido (Mapa 3 [(Rio 2, [Nenhum,Tronco,Tronco]),(Rio 2 ,[Nenhum,Tronco,Tronco]),(Relva, [Nenhum,Nenhum,Nenhum])])
False
>>> mapaValido (Mapa 3 [(Estrada 2, [Carro,Carro,Carro]),(Rio -2 ,[Nenhum,Tronco,Tronco]),(Relva, [Nenhum,Nenhum,Nenhum])])
False  

checkTerreno :: Mapa -> Bool

A função checkTerreno verifica se um dado tereno é valido tendo em conta os que o precedem e os obstáculos contidos nea mesma linha

checkRio :: [Obstaculo] -> Int -> Bool

A função checkRio verifica se os obstáculos numa linha com o terreno Rio são válidos

Exemplo de utilização

>>> checkRio [Tronco,Tronco,Tronco,Tronco,Tronco,Tronco,Nenhum] 5
False
>>> checkRio [Tronco,Nenhum,Tronco,Tronco,Tronco] 5
True

checkEstrada :: [Obstaculo] -> Int -> Bool

A função checkEstrada verifica se os obstáculos numa linha com o terreno Estrada são válidos

Exemplo de utilização

>>> checkEstrada [Carro,Nenhum,Carro,Carro,Nenhum] 3
True

checkRelva :: [Obstaculo] -> Bool

A função checkRelva verifica se os obstáculos numa linha com o terreno Relva são válidos

Exemplo de utilização

>>> checkRelva [Arvore,Nenhum,Arvore,Nenhum]
True

checkNenhum :: Mapa -> Bool

A função checkNenhum verifica,linha a linha, se existe pelo menos um obstaculo Nenhum em todas as linhas

Exemplo de utilização

>>> checkNenhum (Mapa 3 [(Relva,[Nenhum,Arvore,Arvore]),(Relva,[Arvore,Arvore,Arvore])])
False

checkLargura :: Largura -> [Obstaculo] -> Bool

A função checkLargura verifica se o comprimento da lista de obstaculos é igual ao valor da largura == Exemplo de utilização

>>> checkLargura 3 [Tronco,Tronco,Nenhum,Tronco]
False

checkDirecoes :: Mapa -> Bool

A função checkDirecoes verifica se linhas contiguas do terreno Rio possuem direções diferentes

Exemplo de utilização

>>> checkDirecoes (Mapa 3 [(Rio (-1),[Tronco,Tronco,Nenhum]),(Rio 1,[Tronco,Tronco,Nenhum])])
True

checkContiguo :: Mapa -> Bool

A função checkContiguo verifica se não existem mais de 4 rios contiguos, ou 5 relvas ou estradas contiguas

Exemplo de utilização

>>> checkContiguo (Mapa 3 [(Rio 1 ,[Tronco,Tronco,Nenhum]),(Relva ,[Arvore,Nenhum,Nenhum]),(Rio 2 ,[Nenhum,Tronco,Nenhum]),(Rio (-3) ,[Tronco,Tronco,Nenhum]),(Rio 1,[Tronco,Tronco,Nenhum])])
True

checkPassavel :: Mapa -> Bool

A função checkPassavel verifica se um dado mapa é passável ou seja pode ser completado sem becos sem fim.

verificaSeIgual :: [Int] -> [Int] -> Bool

A função verificaSeIgual verifica se existes uma coluna livre de obstáculos no mapa para que este seja passável.