Tarefa2_2022li1g091

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2022/23.

Documentation

randomIntsL :: Int -> Int -> [Int]

A função randomIntsL gera uma lista de números aleatórios a partir de uma seed e um valor de comprimento da lista.

rnd :: Int -> Int

A função rnd devolve um número aleatŕoio que será utilisado pelas diversas funções.

estendeMapa :: Mapa -> Int -> Mapa

A função estendeMapa adiciona uma nova linha ao "topo" do mapa.

novaLinha :: Mapa -> Int -> (Terreno, [Obstaculo])

A função novalinha calcula um terreno aleatorio que será utilizado na nova linha.

addVelocidades :: Mapa -> Int -> (Terreno, [Obstaculo])

A função addVelocidades devolve um valor aleatório possível de uma velocidade dependenod do tipo do terreno.

novoObstaculo :: Mapa -> Int -> (Terreno, [Obstaculo])

A função novoObstaculo cria uma lista possível de obstaculos correspondete a um tipo de terreno de uma dada linha.

proximosTerrenosValidos :: Mapa -> [Terreno]

A função proximosTerrenosValidos devolve uma lista dos possíveis terrenos válidos tendo em conta os anteriores no mapa.

Exemplo de utilização

>>> proximosTerrenosValidos (Mapa 3 [(Estrada 1 ,[Carro,Carro,Nenhum]),(Estrada 2 ,[Carro,Carro,Nenhum]),(Estrada 3 ,[Carro,Carro,Nenhum]),(Estrada (-3) ,[Carro,Carro,Nenhum]),(Estrada 1,[Carro,Carro,Nenhum])])
[Rio 0, Relva]

proximosObstaculosValidos :: Int -> (Terreno, [Obstaculo]) -> [Obstaculo]

A função proximosObstaculosValidos devolve uma lista com os possíveis obstaculos validos tendo em conta o terreno e os obstaculos anteriores.

checkRio1 :: Int -> [Obstaculo] -> [Obstaculo]

A função checkRio1 devolve uma lista com os possíveis obstaculos quando o terreno é do tipo rio.

Exemplo de utilização

>>> checkRio1 4 [Tronco,Tronco,Tronco]
[Nenhum]

checkEstrada1 :: Int -> [Obstaculo] -> [Obstaculo]

A função checkEstrada1 devolve uma lista com os possíveis obstaculos quando o terreno é do tipo estrada.

Exemplo de utilização

>>> checkEstrada1 5 [Carro,Carro,Carro]
[Nenhum]

checkRelva1 :: Int -> [Obstaculo] -> [Obstaculo]

A função checkRelva1 devolve uma lista com os possíveis obstaculos quando o terreno é do tipo relva.

Exemplo de utilização

>>> checkRelva1 3 [Nenhum,Arvore]
[Arvore,Nenhum]

checkContiguo1 :: Mapa -> Bool

A função checkContiguo verifica se não existem mais de 3 rios contiguos, ou 4 relvas ou estradas contiguas

Exemplo de utilização

>>> checkContiguo1 (Mapa 3 [(Rio 1 ,[Tronco,Tronco,Nenhum]),(Rio (-2) ,[Tronco,Nenhum,Nenhum]),(Rio 2 ,[Nenhum,Tronco,Nenhum]),(Rio (-3) ,[Tronco,Tronco,Nenhum])])
False