Tarefa3_2022li1g091

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2022/23.

Documentation

animaJogo :: Jogo -> Jogada -> Jogo

A função animaJogo está responsável por animar o mapa a cada momento que passa e o jogador dependendo da sua jogada. As duas linhas que se encontram em forma de comentário é a parte pedida na segunda fase do projeto em relação aos troncos. Estas estão desativadas no momento para melhorar a jogabilidade do jogo em si mas se for necessário podem ser ativadas outra vez.

Exemplo de utilização

>>> animaJogo (Jogo (Jogador (2,1)) (Mapa 5 [(Relva,[Arvore,Nenhum,Nenhum,Nenhum,Arvore]),(Rio 1,[Nenhum,Nenhum,Tronco,Nenhum,Nenhum])])) (Move Cima)
 (Jogo (Jogador (2,0)) (Mapa 5 [(Relva,[Arvore,Nenhum,Nenhum,Nenhum,Arvore]),(Rio 1,[Nenhum,Nenhum,Nenhum,Tronco,Nenhum])]))
>>> animaJogo (Jogo (Jogador (2,1)) (Mapa 5 [(Relva,[Arvore,Nenhum,Arvore,Nenhum,Arvore]),(Rio 1,[Nenhum,Nenhum,Tronco,Nenhum,Nenhum])])) Parado
 (Jogo (Jogador (3,1)) (Mapa 5 [(Relva,[Arvore,Nenhum,Nenhum,Nenhum,Arvore]),(Rio 1,[Nenhum,Nenhum,Nenhum,Tronco,Nenhum])]))

isRio :: Terreno -> Bool

A função isRio verifica se uma linha é um rio.

speed :: Terreno -> Velocidade

A função speed devolve a velocidade de um terreno sem o tipo (Relva,Rio ou Estrada).

speedRio :: Jogo -> Int -> Velocidade

A função speedRio devolve a velocidade do rio especifico em que o jogador se encontra.

Exemplo de utilização

>>> speedRio (Jogo (Jogador (2,1)) (Mapa 5 [(Relva,[Arvore,Nenhum,Arvore,Nenhum,Arvore]),(Rio 1,[Nenhum,Nenhum,Tronco,Nenhum,Nenhum])])) 0
 1

linhatronco :: Jogo -> Int -> Bool

A função linhatronco verifica se o Jogador está em cima de um tronco.

Exemplo de utilização

>>> linhatronco (Jogo (Jogador (2,1)) (Mapa 5 [(Relva,[Arvore,Nenhum,Arvore,Nenhum,Arvore]),(Rio 1,[Nenhum,Nenhum,Tronco,Nenhum,Nenhum])])) 0
 True

moveplayer :: Jogo -> Jogada -> Jogador

A função moveplayer move o jogador dependendo da jogada que faz.

Exemplo de utilização

>>> moveplayer (Jogo (Jogador (2,0)) (Mapa 5 [(Relva,[Arvore,Nenhum,Nenhum,Nenhum,Arvore]),(Rio 1,[Nenhum,Nenhum,Tronco,Nenhum,Nenhum])])) (Move Esquerda)
 Jogador (1,0)

linhaArvore :: Jogo -> Int -> Bool

A função linhaArvore verifica se o Jogador está na mesma posição que uma árvore.

existeArvore :: Jogo -> Jogada -> Bool

A função existeArvore impede o Jogador de passar por árvores como se fosse um fantasma.

isEstrada :: Terreno -> Bool

A função isEstrada verifica se uma linha é uma estrada.

linhaCarro :: Jogo -> Int -> Bool

A função linhaCarro verifica se o Jogador está na mesma posição que um carro.

atropelado :: Jogo -> Jogada -> Bool

A função atropelado verifica se o jogador ao mover-se encontra-se na mesma posição que um carro.

mexelistaupdt :: Terreno -> [Obstaculo] -> Jogo -> Jogada -> [Obstaculo]

A função mexelistaupdt move todos os elementos de uma lista um certo valor para a frente ou para trás e tambem essa lista de avançar se a qualquer ponto o jogador estiver na mesma posição que um carro.

Exemplo de utilização

>>> mexelistaupdt (Rio 1) [Nenhum,Nenhum,Tronco,Nenhum,Nenhum] (Jogo(Jogador (2,1)) (Mapa 5 [(Relva,[Arvore,Nenhum,Nenhum,Nenhum,Arvore]),(Rio 1,[Nenhum,Nenhum,Tronco,Nenhum,Nenhum])])) (Parado)
[Nenhum,Nenhum,Nenhum,Tronco,Nenhum]

mexecarrooutroncoupdt :: Jogo -> Jogada -> [(Terreno, [Obstaculo])]

A função mexecarrooutroncoupdt faz com que o mapa seja atualizado independentemente do Jogador, ou seja, mexe os carros e os troncos de acordo com a velocidade do respetivo terreno.

Exemplo de utilização

>>> mexecarrooutroncoupdt (Jogo(Jogador (2,1)) (Mapa 5 [(Relva,[Arvore,Nenhum,Arvore,Nenhum,Arvore]),(Rio 1,[Nenhum,Nenhum,Tronco,Nenhum,Nenhum])])) (Parado)
 [(Relva,[Arvore,Nenhum,Arvore,Nenhum,Arvore]),(Rio 1,[Nenhum,Nenhum,Nenhum,Tronco,Nenhum])]