Tarefa4_2022li1g091

Description

Módulo para a realização da Tarefa 4 do projeto de LI1 em 2022/23.

Documentation

jogoTerminou :: Jogo -> Bool

A função jogoTerminou vai confirmar se o Jogador está morto ou se está vivo, ou seja, se o jogador se encontrar fora do Mapa, se estiver na água ou se for atropelado então o Jogador está morto e o jogo acaba.

Exemplo de utilização

>>> jogoTerminou (Jogo (Jogador (2,1)) (Mapa 5 [(Relva,[Arvore,Nenhum,Nenhum,Nenhum,Arvore]),(Rio 1,[Nenhum,Nenhum,Tronco,Nenhum,Nenhum])]))
 False
>>> jogoTerminou (Jogo (Jogador (1,1)) (Mapa 5 [(Relva,[Arvore,Nenhum,Arvore,Nenhum,Arvore]),(Rio 1,[Nenhum,Nenhum,Tronco,Nenhum,Nenhum])]))
 True

linhaCarro :: Jogo -> Int -> Bool

A função linhaCarro verifica se o Jogador está na mesma posição que um carro, ou seja, se foi atropelado por um carro.

Exemplo de utilização

>>> linhaCarro (Jogo (Jogador (2,1)) (Mapa 5 [(Relva,[Arvore,Nenhum,Arvore,Nenhum,Arvore]),(Estrada 1,[Nenhum,Nenhum,Carro,Nenhum,Nenhum])])) 0
 True

linhaagua :: Jogo -> Int -> Bool

A função linhaagua verifica se o Jogador não está em cima de um tronco quando está num rio.

Exemplo de utilização

>>> linhaagua (Jogo (Jogador (1,1)) (Mapa 5 [(Relva,[Arvore,Nenhum,Arvore,Nenhum,Arvore]),(Rio 1,[Nenhum,Nenhum,Tronco,Nenhum,Nenhum])])) 0
 True

isRio :: Terreno -> Bool

A função isRio verifica se uma linha é um rio.

isEstrada :: Terreno -> Bool

A função isEstrada verifica se uma linha é uma estrada.